Move marker operation to math mod
