+++ src/math/mod.rs
@@ -2,7 +2,11 @@
 // Use of this source code is governed by a MIT-style
 // license that can be found in the LICENSE file.
 
-use std::ops::Deref;
+use std::ops::{
+    BitXor,
+    Deref,
+};
+use vector::Vector;
 
 pub use self::mat::Mat;
 
@@ -30,4 +34,16 @@ impl<A> Deref for Trans<A> {
 pub enum Marker {
     T,
     H,
+}
+
+impl<'a, T> BitXor<Marker> for &'a Vector<T>
+{
+    type Output = Trans<&'a Vector<T>>;
+
+    fn bitxor(self, m: Marker) -> Trans<&'a Vector<T>> {
+        match m {
+            Marker::T => Trans::T(self),
+            Marker::H => Trans::H(self),
+        }
+    }
 }
