+++ src/math/vector.rs
@@ -4,29 +4,13 @@
 
 use std::ops::{
     Add,
-    BitXor,
     Mul,
 };
 use num::complex::{Complex32, Complex64};
 use default::Default;
 use vector::ops::*;
 use vector::Vector;
-use math::{
-    Marker,
-    Trans,
-};
-
-impl<'a, T> BitXor<Marker> for &'a Vector<T>
-{
-    type Output = Trans<&'a Vector<T>>;
-
-    fn bitxor(self, m: Marker) -> Trans<&'a Vector<T>> {
-        match m {
-            Marker::T => Trans::T(self),
-            Marker::H => Trans::H(self),
-        }
-    }
-}
+use math::Trans;
 
 impl<'a, T> Add for &'a Vector<T>
     where T: Axpy + Copy + Default
